CLN: big refactor of intrinsic_sampling_method module
