+++ alias/io/checkfile_io.py
@@ -2,47 +2,47 @@ import pickle
 
 
 def make_checkfile(checkfile_name):
-	"""
-	Creates checkfile for analysis, storing key paramtere
-	"""
+    """
+    Creates checkfile for analysis, storing key paramtere
+    """
 
-	checkfile = {}
-	with open(checkfile_name + '.pkl', 'wb') as outfile:
-		pickle.dump(checkfile, outfile, pickle.HIGHEST_PROTOCOL)
+    checkfile = {}
+    with open(checkfile_name + '.pkl', 'wb') as outfile:
+        pickle.dump(checkfile, outfile, pickle.HIGHEST_PROTOCOL)
 
 
 def read_checkfile(checkfile_name):
-	"""
-	Reads checkfile to lookup stored key paramters
-	"""
+    """
+    Reads checkfile to lookup stored key paramters
+    """
 
-	with open(checkfile_name + '.pkl', 'rb') as infile:
-		return pickle.load(infile)
+    with open(checkfile_name + '.pkl', 'rb') as infile:
+        return pickle.load(infile)
 
 
 def update_checkfile(checkfile_name, symb, obj):
-	"""
-	Updates checkfile parameter
+    """
+    Updates checkfile parameter
 
-	Parameters
-	----------
-	checkfile_name:  str
-			Checkfile path + name
-	symb:  str
-			Key for checkfile dictionary of object obj
-	obj:
-			Parameter to be saved
+    Parameters
+    ----------
+    checkfile_name:  str
+            Checkfile path + name
+    symb:  str
+            Key for checkfile dictionary of object obj
+    obj:
+            Parameter to be saved
 
-	Returns
-	-------
-	checkfile:  dict
-			Dictionary of key parameters
-	"""
+    Returns
+    -------
+    checkfile:  dict
+            Dictionary of key parameters
+    """
 
-	with open(checkfile_name + '.pkl', 'rb') as infile:
-		checkfile = pickle.load(infile)
-	checkfile[symb] = obj
+    with open(checkfile_name + '.pkl', 'rb') as infile:
+        checkfile = pickle.load(infile)
+    checkfile[symb] = obj
 
-	with open(checkfile_name + '.pkl', 'wb') as outfile:
-		pickle.dump(checkfile, outfile, pickle.HIGHEST_PROTOCOL)
-	return checkfile
+    with open(checkfile_name + '.pkl', 'wb') as outfile:
+        pickle.dump(checkfile, outfile, pickle.HIGHEST_PROTOCOL)
+    return checkfile
